pyproject.toml
[tool.poetry]
name = "rlc_task_solution"
version = "0.1.0"
description = ""
authors = ["kypi-mne-tapo4ki"]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.109.0"
uvicorn = "^0.27.0"
pydantic = "^2.6.0"
pyyaml = "^6.0.1"
PyJWT = "^2.8.0"
alembic = "^1.13.1"
asyncpg = "^0.29.0"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.25"}

[tool.poetry.group.dev.dependencies]
httpx = "^0.26.0"
pytest = "^7.4.0"
black = "^24.1.1"
flake8 = "^7.0.0"

[build-system]
requires = ["poetry-core>=1.5.1"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.black]
line-length = 79
